refactor(mm): move MemoryManagement to its own file
